Link users to todos.

+++ convex/functions.ts
@@ -1,8 +1,14 @@
 import { mutation, query } from "./_generated/server";
 import { v } from "convex/values";
+import {requireUser} from "./helper"
+
 export const listTodos = query(
     {handler: async (ctx) => {
-        return await ctx.db.query("todos").collect();
+        const user = await requireUser(ctx);
+        return await ctx.db.query("todos")
+        .withIndex("by_user_id", q=>q.eq("userID",user.tokenIdentifier))
+        // .filter(q=>q.eq(q.field("userID"),user.tokenIdentifier))
+        .collect();
     }}
 )
 
@@ -12,10 +18,13 @@ export const createTodo = mutation({
         description: v.string()
     },
     handler: async (ctx, args) => {
-        return await ctx.db.insert("todos", {
+        const user = await requireUser(ctx);
+        await ctx.db.insert("todos", {
              title: args.title, 
              description: args.description, 
-             completed: false });
+             completed: false, 
+             userID:user.tokenIdentifier
+            });
     },
 });
 export const updateTodo = mutation({
@@ -24,6 +33,11 @@ export const updateTodo = mutation({
         completed: v.boolean(),
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id)
+        if(todo?.userID != user.tokenIdentifier){
+            throw new Error("Unauthorized");
+        }
          await ctx.db.patch(args.id,{
              completed: args.completed 
             });
@@ -34,7 +48,11 @@ export const deleteTodo = mutation({
         id: v.id("todos"),
     },
     handler: async (ctx, args) => {
-        console.log(ctx.db);
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id)
+        if(todo?.userID != user.tokenIdentifier){
+            throw new Error("Unauthorized");
+        }
         await ctx.db.delete(args.id);
     },
 });
